+++ app/models/maths_and_physics/slug_sequence.rb
@@ -41,6 +41,7 @@ module MathsAndPhysics
       "email-verification",
       "provide-mobile-number",
       "mobile-number",
+      "mobile-verification",
       "bank-or-building-society",
       "personal-bank-account",
       "building-society-account",