+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog]
 
 - All policies
   - Prevent non-selection of an address option on the home address page
+  - Add mobile verification with OTP to TSLR & M+P policies
 
 ## [Release 104] - 2021-11-25